Added library code for MultipartRequests (to upload pictures)
Added permissions to manifest
Also changed the target sdk to Lollipop (API 22) as this removed a permission error for uploading posts.

Uploading posts is now working, although there is an issue with getting the picture you have just taken as they dont appear in the gallery until the emulated sd card is remounted. This can be avoided by using on a real phone apparently.